Queue booking created task if trigger is enabled
